Create Zomato Dataset-EDA
Bengaluru is the IT capital of India. Most of the people here are dependent mainly on restaurant food as they don’t have time to cook for themselves.
With such an overwhelming demand for restaurants, it has therefore become essential to study the demography of a location. What kind of food is more prevalent in a locality? Do the entire locality loves vegetarian food? If yes then is that locality populated by a particular sect of people for eg. Jain, Marwaris, Gujaratis who are
mostly vegetarian. This kind of analysis can be done using the data, by studying factors such as
• Location of the restaurant
• Approx Price of food
• Theme-based restaurant or not
• Which locality of that city serves that cuisine with the maximum number of restaurants
• The facilities like online orders or book tables.
• Is a particular neighborhood famous for its own kind of food.